cpu client: major cleanup
The original cpu implementation followed the plan 9 design, wherein
all possible ways of using it were contained in one binary.

With the cpud split out for reasons of space and limitation
of blast radius, it's become too confusing to try to keep all the
code in sync.

Until this is redone with server/ and client/ packages, remove
all vestiges of init and server support from the cpu command.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>