more cleanup
redo timeout code, to use time.After. Add comment about
weird issues with the timeout.

Overall, it is a lot more solid with, e.g., this loop
for i in /usr/bin/* ; do ./cpu localhost date; done

it is not perfect, and sometimes we have seen a segv
in the ssh package.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>