cpud and cpu: improve debugging, fix unshare bug
It used to be, AFAIK, safe to call Unshare() from init().
Or maybe not? There were persistent failures in setting up
the namespace, so maybe this has always not worked.
It would fail several times per hundred iterations,
which was far too many.

This is basically a go runtime problem
which has always existed, but the Go team has never
felt motivated to discuss fixes (not surprising; Unshare
is a real corner case with little or no use). I've even
had a proposed CL for a fix languish

Also, put a capability in cpud to drop to a script when things
fail. It is controlled by the wtf option. One can temporarily
hardwire it in the code or invoke cpud with, e.g.,
-wtf /bbin/elvish
added to the switches. In the event of a namespace
failure, wtf will run, and it can be a shell or just
a command such as mount.

Have all messages from cpud prefixed with CPUD:
it's far easier to find.

With these changes, thousands of runs of cpu exhibit no
namespace issues.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>