cpud: one more go at fixing orphans
The process management was wrong. Basically, how it must be done:

cpud, in init, starts a process reaper which can never exit and must
run as PID 1. Use LockOSThread to ensure that, and if wait4 returns
empty-handed, sleep one second and try again.

For cpud as -remote, ONLY wait for the process it started, for reasons
described in the code.

This change fixes a lot of corner cases seen in the wild.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>