cpu: correctly handle exit status
Until now, due to incorrect exit status handling in cpud,
exit status was not correctly returned to the client.

Now that cpud is fixed, cpu, on error, gets the exit status from
the ssh.ExitStatus if that is the return type; or it will use
1, and do a deferr'ed os.Exit call in main(). The defer is so that
tty modes can be restored.

This fixes scripts that invoked cpu and did not work right as they
never got an error.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>