cmds/cpud: allow flexible namespace remapping
The CPU_NAMESPACE variable now looks like this:
CPU_NAMESPACE=[path[=path][:path[=path]]*]

It can be empty:
CPU_NAMESPACE=
in which case no 9p mount is needed or tried;
or it can have a set of paths, as today:
CPU_NAMESPACE=/bin:/usr/bin

such that /bin on the remote is mounted over /bin on the local system;

But not all systems have identical paths. There is no /home on OSX,
but there is /Users. So we need a way to say:
mount the remote /Users/rminnich on /home/rminnich.

We can do this now:

CPU_NAMESPACE=/home/rminnich=/Users/rminnich

so that we can map different paths from a remote to a local.

TODO: it would be nice to have a way to say "add this extra
stuff to the default namespace."

Would that be CPU_EXTRANAMESPACE?

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>