cpud: return no error if the wait returns ECHILD
It is clear that the wait behavior needs more work, however.

The situation is that the init process starts itself as runRemote,
and when the remote runner starts and runs the child, it seems
our zombie grabber snarfs the wait status.

This only is happening on the pcengines apu2, if you can believe it.

This is on me, I clearly got this wrong (again).

For whatever reason, returning ECHILD over the wire
is not something we do?

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>