Major package surgery
Work on contest revealed how bad the design was.
Here begin the fixes.

cpu/: implement a test which repros contest failures

Adding With form to see if it's a bit nicer looking to use
c := Command(host, ...args).WithRoot("/").WithPort(xx)
etc.

It has proven to be nice to use in contest

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>