Ignore shell errors
The return code for bash is the return code for the last command in
bash.  If cpu HOST bash then ran a command that failed, cpud would think
that the shell failed.  It wouldn't propagate the exact error code
either.  cpu on the client side would get an exit code of 1.  e.g.

	$ cpu HOST sh
	sh-5.1# ^C
	sh-5.1#
	exit
	2022/03/10 08:55:28 CPUD(as remote):exit status 130
	2022/03/10 08:55:28 SSH error Process exited with status 1
	$ echo $?
	1

Although it might be nice to propagate the error code (e.g. 130) to cpu,
just ignoring the error is probably easier for users to handle.

sshd seems to have some error propagation.  If you ctl-c, then ctl-d,
ssh returns 0.  If you ls no-such-file, ssh returns 1, not 2 (which is
what bash returns).

Alternatively we could just propagate the '1', like ssh does, and not
spit out the error messages.

Signed-off-by: Barret Rhoden <brho@google.com>