Add vsock support to cpud and cpu client package
vsock is a simpler networking path for virtual machine guests.

It needs no emulated hardware or IP stacks. It is a lot like
networking over a serial line. There are fewer moving parts
and hence less to go wrong.

One thing that can go wrong is your kernel is broken, e.g. ubuntu
5.8, and in that case you will get connection timeouts. To fix
a problem like this, update ubuntu or use something else. Ubuntu
22 is known to work; ubuntu 20 is known to not work.

Sadly, vsock is not in the net package, so the code is more complex
than it ought to be.

To use vsock on cpud:
cpud -net vsock [... args]

cpud will listen on any CID, so can be reached even from the VM itself.

To use it on the client:
./cpu -net vsock [other switches] vsock-client-id [other cpu args]

e.g., if one starts qemu with the following vsock switches:
-device vhost-vsock-pci,guest-cid=42

The vsock-client-id above would be 42.

For example, when qemu is started as above, then the cpud command is
cpud -net vsock -init

and the cpu command can be (assuming you are not using a default
key):
./cpu -key ~/.ssh/cpu_rsa -net vsock 42 date

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>