Add an "inetd" mode for cpud
People are looking at running cpud manually, after the system is fully booted.

cpud will now have a default behavior when it is invoked with no mode: it will
run as "inetd." It will not try to set up cgroups, loopback interfaces, or a
process reaper: that is for init to do.

This will make it easier to run cpud on systems which are past runlevel 1.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>