Use /proc/self/exe on Linux for cpud
By specifying the command "cpud -remote", we're hoping that cpud is in
the PATH of cpud itself.  If it's in a one-off directory, it'll fail.
The client could set -cpudcmd to specify the full path to cpud, but
there's no reason for the client to know where cpud is.

On Linux, /proc/self/exe, when run in the context of cpud, will point to
cpud itself.  This is a lot safer than -cpudcmd btw, so perhaps we can
remove that option in the future.

Signed-off-by: Barret Rhoden <brho@google.com>