Add the dut command from linuxboot
It is much easier to bring dut into cpu than have 2 repos.
Further, dut is not a combined cpu and cpud; it is useful
to have this illustrative use of the cpu package.

Also, dut is changed so that it will be able to work with
different kinds of networks, although vsock support
will likely require special handling.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>