dnssd: add sorting to the client
You can now sort on various meta-data.  Default is to sort numerically
ascending, but you can override with comparitors (<, >, !, or =).
In the case of ! of = they will sort based on the existence (or
non-existence of a key).

Since we are now picking amongst a list, there is a built in delay of 1
second to gather results before sorting.

e.g. /bin/cpu dnssd:?sort=order\&sort==eorder\&sort=\<gtorder

Signed-off-by: Eric Van Hensbergen <eric.vanhensbergen@arm.com>