client: only pass CPU_FSTAB to cpud
cpu has been able to pass an fstab in CPU_FSTAB for a while.

This led to having two ways to specify a namespace to cpud:
CPU_NAMESPACE and CPU_FSTAB.

We can simplify this, and simplify cpud, by having the client
parse the namespace and add it to CPU_FSTAB.

Further, if the client parses the contents of -namespace
for errors, feedback for bad values is much more
immediate.

This change:
o implements parsing the namespace to CPU_FSTAB
o concatenating it at the end of a user-specified
  fstab file (if any)
o NOT passing CPU_NAMESPACE to cpud

If this works, the next step is to remove CPU_NAMESPACE support
from cpud, thereby simplifying it.

The namespace switch will be preserved for the client, because
it is very convenient.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>