Refactor dns-sd into its own module and cmds
This patch refactors dns-sd to seperate it a bit more from
mainline cpu.

The dns-sd code is now in a ds module.  There is now a dcpu
and dcpud command line which incorporate the dns-sd functionality.

I've also added a docker-based integration test in TESTDOCKERCPU
which tests both the original cpu/cpud and tests dcpu/dcpud.

Signed-off-by: Eric Van Hensbergen <eric.vanhensbergen@arm.com>