Update documentation for docker
Point people are the ghcr.io container which is now automatically
populated by gitlab CI.  The container version is pegged to the branch
at the moment so its cpu:main, I wasn't sure how you wanted to handle
release flow of u-root/cpu:latest, probably best for now to create a
custom action that is only triggerable by an owner to tag whatever is in
cpu:main to cpu:latest.  Once we've done that we can update docs again.

Signed-off-by: Eric Van Hensbergen <ericvh@gmail.com>