client.go: Set std in/out/err like exec.Cmd
The Stdin/Stdout/Stderr fields of exec.Cmd represents the file
descriptors that the cmd can use as its stdin/stdout. So Stdin of
exec.Cmd is an io.Reader and Stdout/Stderr are io.Writer.

On the other hand, the current Stdin/Stdout/Stderr fields of
client.Cmd have different meanings: they are actually the pipes
for fetching data from or writing data to the SSH session, so Stdin
is a writer and stdout/stderr are readers.

This PR renames the original Stdin/Stdout/Stderr to
SessionIn/SessionOut/SessionErr and adds new Stdin/Stdout/Stderr
fields similar to exec.Cmd.

Signed-off-by: Changyuan Lyu <changyuanl@google.com>