cpu.go: Make sure c.Close() is called on all paths
Currently if c.Dial() or c.Start() fails, the closers are not called.
Also, if c.Wait() fails, the corresponding error is dropped and not
returned back to caller. As a result, in most cases, if a command
failed on the remote server, the cpu process still exits with 0.

This PR makes sure c.Close() is called on all paths and the error is
returned to the caller. It also makes sure closers are called in a
last-in-first-called order.

Signed-off-by: Changyuan Lyu <changyuanl@google.com>