client: escape command arguments
Currently, If we run
  ./cpu <host-name> ./binary --arg1 "value with space"
The client wiil actually run the following command on the remote host
  ./binary --arg1 value with space
where arg1=value and the binary gets two extra arguments `with` and
`space`, because the quote is lost.

This PR imports the shellescape package. With the fix, the remote
host will receive a correctly quoted command line:
  ./binary --arg1 'value with space'

Signed-off-by: Changyuan Lyu <changyuanl@google.com>