Fix verbose printing
This PR fix the usage of verbose() and v(). v() will be the log
function provided by the user and verbose() is used in the package.
v() is not called directly besides verbose().

We need to consider 3 kinds of verbose printings:
1. printings from cpud ssh server
2. printings from cpud remote mode
3. printings from local cpu command

The 1st will be logged to server, so we do not need to care about
terminal interface. For the 2nd and 3rd cases, logs will be printed or
forwarded to local terminal window. Since local terminal will also be
used as the ssh terminal window, we need to add the extra `\r\n` to
the end of log messages.

Signed-off-by: Changyuan Lyu <changyuanl@google.com>