cpud: add a switch to 'register' cpud with a remote host
In VMs and IoT, it is useful to have some idea that a host
has come up. Pinging the cpud port until cpud comes up does not work
well, especially in systems using vsock. What is most desired
is that cpud 'check in', or register, with a control host
so that the contol host can keep track of which cpuds are up
and working.

Once the cpud has registered, the control host can periodically
run a command to ensure the host is still up.

This PR adds a flag (default off) which tells cpud to
register, what address to use, and a second flag to
tell cpud how long to wait for the register to succeed.

The register flag is a Go-style addr (host:port) which specifies a
host to connect to once cpud has called Listen and before
it calls Serve. cpud will use the network on which it listens.
Assuming the connect succeeds, cpud will send the two-byte
string 'ok' to indicate its status.

registerTO specifies the timeout in seconds for the connect.

Currently, a failure to register or write 'ok' does not block
cpud from continuing on to accept connections. Further, cpud will
only print an error message if debug is enabled.

This PR includes a test for the new function.

Sample usage:

A remote host, atomicpi, is running cpud. Our localhost is
192.168.0.1
To test the register switch, we can cpu to atomicpi, and start
a second cpud running on a different port, and test the register
switch. On the local host, we can use nc (netcat) to wait for the
cpud to register.

On the local host, we first start a netcat:
rminnich@a300:~/linuxboot/mainboards/digitalloggers/atomicpi$ nc -l 4040

In another window, we run a cpu command to start a cpu daemon on
a different service port. We tell the cpud to register to the local host
(192.168.0.1) and port 4040, which our nc command is listening:

cpu atomicpi ./cpud -sp 8086 -register 192.168.0.1:4040 -d
2012/06/16 03:58:07 cpud: running as a server (a.k.a. starter of cpud's for sessions
2012/06/16 03:58:07 configure SSH server
2012/06/16 03:58:07 Server is &{:17010 0x5de880 []  <nil> <nil> 0x5df860 <nil> <nil> 0x5df3a0 0x5df480 <nil> <nil> <nil> 0s 0s map[] map[cancel-tcpip-forward:0x5dfa00 tcpip-forward:0x5dfa00] map[] {{} 0 0} {{0 0} 0 0 0 0} map[] map[] {{} 0 0} <nil>}
2012/06/16 03:58:07 Listening on 0.0.0.0:8086

The cpud has started.
Further, since we used the -register switch, it will have tried to connect
to port 192.168.0.1:4040, and send the string 'ok'.

In the other window, we can see the netcat return

okrminnich@a300:~/linuxboot/mainboards/digitalloggers/atomicpi$

The 'ok' is from cpud.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>