session:Use ip6 and ip4 loopback address, not "localhost"
We had moved to using the name "localhost", since at some
sites, only one of ip6 or ip4 protocols are provided, and,
hence, using the name localhost allowed setting a preference.

cpu should be able to function with missing or damaged host
files, however.

Change the Dial code in session to use the two well known
ip loopback address, ::1, and 127.0.0.1, in that order.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>