Have the cpud specify which cpud to run
The server package was blindly specifying "cpud", which
is clearly a mistake.

A proposed option was to use os.Args[0], which is concerning
as programs might change os.Args (for, e.g., security reasons)
and having a package blindly depend on that global is similary
worrisome.

Change server.New to accept a third argument, then path of the
cpud; and use a closure in the ssh handler so that argument
can be used by the ssh package.

Signed-off-by: Ronald G. Minnich <rminnich@google.com>