cmds/cpu, cmds/decpu: set c.Env to os.Environ
c.Env is not longer automatically set in the Cmd,
so the caller must now set it.

This has been tested.

cpu to a host works as before

env -i cpu to the host fails as many environment variables are missing.

Hence, the package is no longer automatically setting variables
from os.Environ()

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>