"revert" 4781f0ec352f4f18c86373f20610ae452bc2f95a (allow setting tmpMnt)
That commit reads as follows:
    Add -tmpMnt to change namespace mount point

    Some applications may need to access the original contents in /tmp
    directory of the local host. Adding -tmpMnt to allow changing the
    private namespace mount point to avoid conflict.

There are a number of problems with this idea, especially as move
to Windows support. Further, it did not really prove to be useful.

This is not exactly a revert, because reverting that commit
was no longer simple, as we have more programs (decpu, decpud)
than we used to.

From here on out, the server's private mounts will be in
only one place: os.TempDir()

Signed-off-by: Ronald G Minnich <rminnich@gmail.com>