client: add a union 9p server
The union9p server gathers provides a single 9p mount for
several 9p servers.

It works much as the Plan 9 namespaces work. Callers specify
a set of names. On Walk, the names are matched against the
names in the Walk, and the first match is used for the Walk.

As in Plan 9, the entire Walk request is sent to the server
once a match hits, not a portion.
i.e., if a the Walk is to "home", "rminnich", and there
is an entry for
home
Then the matching server will receive "home", "rminnich", not just
"rminnich"

Signed-off-by: Ronald G Minnich <rminnich@gmail.com>