extend the cpud server banner to include GOOS and GOARCH
The banner now looks like this:
SSH-2.0-linux amd64

Should it be changed/extended?

This information is useful to sidecore clients which
need to pick a flattened Docker cpio image.

It is likely to be useful in general for other
cpu usage, e.g., picking a $PATH for binaries, as
in Plan 9, with its /$ARCH/bin directory structure.

Signed-off-by: Ron Minnich <rminnich@gmail.com>