Cpud double check args (#246)
* ds: improve the message when a service is not found

Show the service that was requested. In the case of default,
users need to see the name.

Signed-off-by: Ronald G Minnich <rminnich@gmail.com>

* cmds/cpud: clean up handling of -remote

If the first arg is -remote=true or -remote,
this is a remote session. At that point, no
other settings of -remote in the args will have
any effect.

This change is necessary as pathological commands
such as
cpud -remote -d -remote=false
will cause incorrect behavior.

The command line options that can now be set from the cpu
command are:
-remote
-port9p
-d

The only one that would be sensible for a user to set
from the command line:
-d

Users wishing to get debug output can, e.g.:
cpu -d hostname command

Signed-off-by: Ronald G Minnich <rminnich@gmail.com>

---------

Signed-off-by: Ronald G Minnich <rminnich@gmail.com>