client: don't set 9p when namespace is non-empty
With the addition of nfs support, we no longer should
assume 9p should be set too, at least not in the
client package.

While this is a change for users of the package, the only
user at present is sidecore, and this actually makes life easier.
In actual use, this change causes no problems.

Signed-off-by: Ronald G Minnich <rminnich@gmail.com>