Nfs3 (#270)
* cpio9p and cpu9p: linting fixup

Signed-off-by: Ronald G Minnich <rminnich@gmail.com>

* client/fns: export JoinFSTab and fix leading newline handling

Signed-off-by: Ronald G Minnich <rminnich@gmail.com>

* client: export ParseBinds

It will be used in the cpu and sidecore commands.

Signed-off-by: Ronald G Minnich <rminnich@gmail.com>

* client: don't set 9p when namespace is non-empty

With the addition of nfs support, we no longer should
assume 9p should be set too, at least not in the
client package.

While this is a change for users of the package, the only
user at present is sidecore, and this actually makes life easier.
In actual use, this change causes no problems.

Signed-off-by: Ronald G Minnich <rminnich@gmail.com>

* cmds/cpu: make the "running" verbose message more useful

Might as well know what host, and what port; particularly
useful when debugging IP6 link layer connections.

Signed-off-by: Ronald G Minnich <rminnich@gmail.com>

* cpu and client: add nfs3 support

nfs3 is more available than 9p; many IoT kernels that support
nfs3 do not have 9p. Further, nfs3 is far faster for common
operations than 9p: in some cases, 100 times faster, and
for bash shell completion, up to 1000 times faster (that
is not a typo).

Add nfs3 support. To enable nfs mounts, invoke cpu with
-9p=false -nfs=true

This change required a significant behavioral change:
even if the namespace is non-empty, 9p will not default
to true.

Signed-off-by: Ronald G Minnich <rminnich@gmail.com>

---------

Signed-off-by: Ronald G Minnich <rminnich@gmail.com>