nfs: remove the cacheHelper provided by go-nfs
There's something not quite right about it, and
it is entirely too ready to return ESTALE,
which does not end well. ESTALE NFS returns
on a shared library are a fastpath to a
segmentation violation.

Even with this removed, the entire go toolchain build
is done in 47s, as compared to 71s with 9p.

Signed-off-by: Ronald G Minnich <rminnich@gmail.com>