Add a cpuns command
This command lets you ssh to a system with the cpu command.
You can then run this command, with or without arguments,
and it will give you a cpu environment.

A typical way to run it:
sudo -E unshare -m cpuns

The -E is important to get the SUDO_UID and SUDO_GID

Signed-off-by: Ronald G Minnich <rminnich@gmail.com>