cmds/cpu: set *ninep and *srvnfs in main() for port 22
If the port is 22, then set ninep and srvnfs
for that case.

Because, going back 8 years, with ninep, we send the Nonce
on the ssh channel, we can not use ninep with sshd.

Sorry about that, but 9p has never gotten to the point that
it is worth fixing this and breaking all current users out there.

Just set things for nfs if we are on port 22.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>