cmds/cpuns: accept args, and pass them on
For the part of cpuns that runs as root, the flags are
not evaluated, save for the debug option. That can be turned off
it if is a concern.

This now works:
~/go/bin/cpu -namespace=/Users s76 bin/cpuns ls /Users
[sudo] password for rminnich:
2024/07/12 17:58:52 [ERROR] No handler for 100227.0
nancyminnich  rminnich	Shared

For .bashrc, one could test the existence of LC_GLENDA_CPU_FSTAB,
and run cpuns automatically.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>