cmds/cpu: -ssh switch; pass environment in command line
if -ssh is true then assume ssh; this is in addition
to the port 22 test.

If on ssh, insert a cpuns command, with a -env switch,
to pass environment variables through all the filters
(sshd and sudo) that are stripping them out.

The only environment variable that is processed is CPU_FSTAB;
the rest are ignored.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>