cpuns: add -env for environment handling
the -env is for cpuns to pass environment to the
eventual spawned process.

It is handled as follows:
If cpuns is not running as root, pass the switch on as -env=

If  cpuns is running as root, break the switch value up with
strings.Split(..., "\n") and append that to the environment.

Signed-off-by: Ronald G Minnich <rminnich@gmail.com>