cpuns: v() -> verbose(),save CPU_FSTAB from sudo
change all v to verbose, so we get the CPUNS: prefix.

some sudo will not blindly accept all environment variables with -E, but
will let us save a select few. CPU_FSTAB seems to be acceptable.

We want to save CPU_FSTAB in the environment so it is not exposed in argv,
as it would be if we added it to the environment variables preserved with
-env, since it exposes the port and the mount GUID.

At this point, with an extremely picky sudo, cpuns works fine.

Signed-off-by: the man <rminnich@google.com>