cpud: clear CPU_FSTAB and LC_GLENDA_CPU_FSTAB before starting command
If you cpu to a machine, and the only command you run is cpu,
the CPU_FSTAB and LC_GLENDA_CPU_FSTAB variables must be cleared,
else the cpu will inherit them and assume it should use them.

Before starting the command, os.Unsetenv CPU_FSTAB and
LC_GLENDA_CPU_FSTAB.

It's changed in both the old code and the new code cpuns uses.

I still have to try to figure out a test.

But, now, on osx, I can cpu into a docker container for a risc-v board,
and from that container cpu into the board.

This is very, very nice. It gets around the problem that docker containers
on osx can not do volume mapping: cpu alone, on osx, has value in and of itself.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>