Add TestIP that uses cpu on 4 Linux architectures
The test runs on Linux on amd64, arm, arm64, and riscv64.

Once this is merged, this test can replace the one in u-root.
Our goal is to elminate as many u-root VM shell script tests
as possible, because we have found undetected errors in them.
Tests should be written in Go.

We keep this test here to also test cpu. Getting it all
working was no small matter. We want it to keep working :-)

We extend the client package in small ways to make writing
VM tests convenient.

The kernel images and initramfs for cpu are built into
the package. The tests can be run hermetically by running
go test.
ghcr.io is only used for go generate, and that will not be
frequent (possibly changing only every year or so).

The kernels are cached from the u-root ghcr.io docker images.
The kernels have a few tweaks, for things missing
from the kernel (GRE, VTI) in ghcr.io. The PR for these additions
has been submitted for the vmtest repo. The kernels also
provide /proc/config.gz. For now, my fixes are local
to config_linux.txt files.

Note: I have no idea how the vmscript in u-root is passing
at this point, with support for GRE, VTI, and other things missing.
The script may not be working; there may be errors
that are not caught. I've seen ignored errors in other scripts.

The compiled-in initramfs package only contains cpud. The vm
package builds a u-root image for testing as needed and concatenates
it to the cpud initramfs.

The VMs start with two network interfaces. The first is used
to support cpud; the second interface can be manipulated by the tests.
This removes the need to parse VM console output, which can be fragile.

IP is configured by ip=dhcp.

The test assumes there is a working qemu on the CI.
This allows us to run these tests on any kernel/architecture
combination that has qemu.

Signed-off-by: Ronald G Minnich <rminnich@gmail.com>